requirements.txt
discord.py==1.4.0
bottle==0.12.17
ipython
